HAP-373 - For multi-module M2 release builds, don't publish artifacts
When releasing a single Maven module with the M2 Release plugin, the
ArtifactoryRedeployPublisher does not publish any artifacts. This was
implemented under HAP-93.

This change makes multi-module Maven builds act the same way. The
artifactory plugin will act as if there is no ArtifactoryRedeployPublisher
configured when performing a M2 Release, and so will not deploy artifacts.

See also JENKINS-16780, which appears to describe the same bug.