[JENKINS-55224] Replace JGit Repository.getRef() call
The JGit project replaced the Repository.getRef(String) method with
Repository.exactRef(String) and Repository.findRef(String) in JGit 5.
The getRef(String) method was removed from JGit 5.2.

JGit compatibility does not provide the same upgrade path as Jenkins
compatibility. The removal of the getRef(String) API will break Jenkins
plugins that depend on git client plugin prior to 3.0. The simplest
solution seems to be to find and replace all calls to Repository.getRef()
with an equivalent call to the Jenkins git client plugin.

Key issues:

* Can't use Repository.getRef() with git client plugin 3.0.
* Can't use Repository.exactRef() with git client plugin before 3.0.

Solution:

* Use Jenkins git client plugin and let it call the correct JGit method